Refactor legacy parsing of provider configuration
Leverage Codable implementation of OpenVPN*.Configuration
